prevent from being able to edit a static course directly through an API call
