+++ api/tests/unit/infrastructure/utils/error-manager_test.js
@@ -1,9 +1,9 @@
 const { expect, hFake } = require ('../../../test-helper');
 const {
   InvalidStaticCourseCreationOrUpdateError,
+  StaticCourseIsInactiveError,
 } = require('../../../../lib/domain/errors');
 const { send } = require('../../../../lib/infrastructure/utils/error-manager');
-//const {} = require('../../../../lib/infrastructure/errors');
 
 describe('Unit | Infrastructure | ErrorManager', function() {
   describe('#send', function() {
@@ -41,6 +41,26 @@ describe('Unit | Infrastructure | ErrorManager', function() {
               pointer: '/data/attributes/challenge-ids',
             },
             detail: ['chalC', 'chalD'],
+          },
+        ],
+      });
+    });
+
+    it('should convert StaticCourseIsInactiveError', async function() {
+      // given
+      const error = new StaticCourseIsInactiveError();
+
+      // when
+      const response = await send(hFake, error);
+
+      // then
+      expect(response.statusCode).to.equal(409);
+      expect(response.source).to.deep.equal({
+        errors: [
+          {
+            detail: 'Opération impossible sur un test statique inactif.',
+            status: '409',
+            title: 'Conflict',
           },
         ],
       });
